moving AbstractToolDefinition
