+++ Makefile
@@ -29,7 +29,8 @@ lint:
 
 .PHONY: typecheck-pyright
 typecheck-pyright:
-	uv run pyright
+	@# PYRIGHT_PYTHON_IGNORE_WARNINGS avoids the overhead of making a request to github on every invocation
+	PYRIGHT_PYTHON_IGNORE_WARNINGS=1 uv run pyright
 
 .PHONY: typecheck-mypy
 typecheck-mypy:
